requirements.txt
# ====================================================================
# Python 依赖包
# 安装命令: pip install -r requirements.txt
# ====================================================================

# 数据获取
akshare>=1.13.0
tushare>=1.3.0
baostock>=0.8.9

# 数据处理
pandas>=2.1.0
numpy>=1.24.0
sqlalchemy>=2.0.0

# LLM
ollama>=0.1.0
requests>=2.31.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 邮件通知
jinja2>=3.1.0

# 可视化
matplotlib>=3.8.0
plotly>=5.18.0
streamlit>=1.29.0

# 工具
tqdm>=4.66.0
schedule>=1.2.0
colorlog>=6.8.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 开发工具
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0